webflux simple version.

+++ src/main/kotlin/com/slipp/race/interface/Race.kt
@@ -1,4 +1,4 @@
 package com.slipp.race.`interface`
 
-class Race(val racers:String = "", val finalLap:Int = 0)
+class Race(val racers:String = "", val finalLap:Long = 0)
